pyproject.toml
[tool.poetry]
name = "chactivity"
version = "1.0.7"
description = "LLM-based Automaic UI Testing Tool"
authors = ["Young-Lord", "66Leo66"]
readme = "README.md"


[tool.poetry.dependencies]
python = "~3.11"
adbutils = "~=1.2.13"
apkutils3 = "~=2.0.11"
appium-python-client = "~=2.11.1"
dacite = "~=1.8.1"
jinja2 = "~=3.1.2"
lxml = "~=4.9.3"
mergedeep = "~=1.3.4"
openai = "~=1.1.1"
pyee = "~=11.0.0"
regex = "~=2023.8.8"
rich = "~=13.5.2"
selenium = "~=4.11.2"
tenacity = "~=8.2.3"
typer = "~=0.9.0"
cachier = "^2.2.1"
types-lxml = "*"
pillow = "^10.1.0"
pylibchecker = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "*"
requests = "*"
pyinstrument = "~=4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
